// src/api/generated/models/purchaseProposalPaymentMethod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type PurchaseProposalPaymentMethod = typeof PurchaseProposalPaymentMethod[keyof typeof PurchaseProposalPaymentMethod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PurchaseProposalPaymentMethod = {
  cash: 'cash',
  financed: 'financed',
  direct_installments: 'direct_installments',
  exchange: 'exchange',
  other: 'other',
} as const;